size-manager: New object to calculate input surface size
The size manager takes the current outputs physical and logical
size into account to calculate height when the corresponding
gsettings are enabled.

This is currently limited to portrait mode.

We target the keyboard's key area height to be 4cm so keys are reasonably
sized. The target height for the dead zone at the bottom is 8mm. This
is useful on taller devices as otherwise when typing the bottom row requires
additional force as the center of gravity is far away from the supporting
fingers.

Signed-off-by: Guido Günther <agx@sigxcpu.org>